update BaseActivity and BaseContract.View
- move showToastMessage method to BaseContract.View
- removed onSaveInstanceStateToBundle abstract method
- changed setUpDataBinding method to getLayoutId method
- move setUpBackArrow method to BaseActivity
